Move AutofillUrlMatcher to impl module
