+++ autofill/autofill-impl/src/main/java/com/duckduckgo/autofill/impl/ui/credential/management/sorting/CredentialInitialExtractor.kt
@@ -17,7 +17,7 @@
 package com.duckduckgo.autofill.impl.ui.credential.management.sorting
 
 import com.duckduckgo.autofill.api.domain.app.LoginCredentials
-import com.duckduckgo.autofill.api.urlmatcher.AutofillUrlMatcher
+import com.duckduckgo.autofill.impl.urlmatcher.AutofillUrlMatcher
 import com.duckduckgo.di.scopes.FragmentScope
 import com.squareup.anvil.annotations.ContributesBinding
 import java.lang.Character.*
