+++ autofill/autofill-impl/src/main/java/com/duckduckgo/autofill/impl/ui/credential/management/sorting/CredentialListSorterByTitleAndDomain.kt
@@ -18,8 +18,8 @@ package com.duckduckgo.autofill.impl.ui.credential.management.sorting
 
 import com.duckduckgo.app.global.extractDomain
 import com.duckduckgo.autofill.api.domain.app.LoginCredentials
-import com.duckduckgo.autofill.api.urlmatcher.AutofillUrlMatcher
-import com.duckduckgo.autofill.api.urlmatcher.AutofillUrlMatcher.ExtractedUrlParts
+import com.duckduckgo.autofill.impl.urlmatcher.AutofillUrlMatcher
+import com.duckduckgo.autofill.impl.urlmatcher.AutofillUrlMatcher.ExtractedUrlParts
 import com.duckduckgo.di.scopes.FragmentScope
 import com.squareup.anvil.annotations.ContributesBinding
 import java.text.Collator
